+++ history.md
@@ -5,7 +5,22 @@
 # Project History
 
 
-2023-11-24: v0.0.12
+v0.0.14
+---------------------------------------------------------------------
+
+- Added Command: $ScanDocuments
+- Added Command: $ProcessDocuments
+- Added Command: $CopyFolder
+- Added Context section of the Tasks file.
+
+
+v0.0.13
+---------------------------------------------------------------------
+
+- Version skipped.
+
+
+v0.0.12 (2023-11-24)
 ---------------------------------------------------------------------
 
 - Added Command: $AppendTextFile